+++ src/version3/models/issueUpdateDetails.ts
@@ -1,3 +1,4 @@
+import { Document } from './document';
 import { EntityProperty } from './entityProperty';
 import { Fields } from './fields';
 import { HistoryMetadata } from './historyMetadata';
@@ -11,7 +12,7 @@ export interface IssueUpdateDetails {
    * provides a straightforward option when setting a sub-field. When multiple sub-fields or other operations are
    * required, use `update`. Fields included in here cannot be included in `update`.
    */
-  fields?: Partial<Fields> | any;
+  fields?: Partial<Omit<Fields, 'description'> & { description: string | Document }> | any;
   /**
    * A Map containing the field field name and a list of operations to perform on the issue screen field. Note that
    * fields included in here cannot be included in `fields`.