+++ src/version3/parameters/updateWorklog.ts
@@ -1,6 +1,6 @@
-import { Worklog } from '../models';
+import { Document, Worklog } from '../models';
 
-export interface UpdateWorklog extends Worklog {
+export interface UpdateWorklog extends Omit<Worklog, 'comment'> {
   /** The ID or key the issue. */
   issueIdOrKey: string;
   /** The ID of the worklog. */
@@ -14,6 +14,12 @@ export interface UpdateWorklog extends Worklog {
    * `auto` Updates the estimate by the difference between the original and updated value of `timeSpent` or `timeSpentSeconds`.
    */
   adjustEstimate?: string;
+  /**
+   * A comment about the worklog in [Atlassian Document
+   * Format](https://developer.atlassian.com/cloud/jira/platform/apis/document/structure/). Optional when creating or
+   * updating a worklog.
+   */
+  comment?: string | Document;
   /**
    * The value to set as the issue's remaining time estimate, as days (#d), hours (#h), or minutes (#m or #). For
    * example, _2d_. Required when `adjustEstimate` is `new`.