+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Jira.js changelog
 
+### 2.15.0
+
+- Version 3: addWorklog and updateWorklog comment field type fixed. Thanks to [RealBuddy](https://github.com/RealBuddy) for reporting the issue.
+- Version 3: maxContentLength was increased for attachments upload. Thanks to [RealBuddy](https://github.com/RealBuddy) for reporting the issue.
+- Version 2, Version 3:
+  - `uiModificationsApps` API added.
+  - `query` property added to `getProjectsForIssueTypeScreenScheme` method `IssueTypeScreenSchemes` API.
+  - `commentId` property added to `getMyPermissions` method `Permissions` API.
+  - `createPriority`, `searchPriorities`, `updatePriority`  methods added to `IssuePriorities` API.
+
 ### 2.14.0
 
 - README: Fast jump to basic authentication topic fixed.